requirements_mlp.txt
absl-py==1.1.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
cloudpickle==1.6.0
cycler==0.11.0
debugpy==1.6.1
decorator==5.1.1
distlib==0.3.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.0.12
flatbuffers==1.12
fonttools==4.33.3
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gym==0.19.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
jsonschema==4.6.2
jupyter-client==7.3.4
jupyter-core==4.11.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.3
kt-legacy==1.0.4
libclang==14.0.1
llvmlite==0.38.1
Markdown==3.3.7
matplotlib==3.5.1
matplotlib-inline==0.1.3
nbformat==5.4.0
nest-asyncio==1.5.5
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.9.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.9.3
shap==0.41.0
six==1.15.0
slicer==0.0.7
stack-data==0.3.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.3.0
urllib3==1.26.9
virtualenv==20.1.0
Werkzeug==2.1.2
xgboost==1.6.0
zipp==3.8.0
